feat: add 404 page on incorrect ingredient url
