+++ src/services/actions/auth.js
@@ -1,7 +1,7 @@
 import { LOG_OUT_URL } from "../../utils/constants";
-import { SIGN_UP_URL, LOG_IN_URL, PROFILE_DATA_URL, accessToken, refreshToken, accessTokenMaxAge, refreshTokenMaxAge, REFRESH_TOKEN_URL, RESET_URL, RECOVER_URL } from "../../utils/constants";
+import { SIGN_UP_URL, LOG_IN_URL, accessToken, refreshToken, accessTokenMaxAge, refreshTokenMaxAge, REFRESH_TOKEN_URL, RESET_URL, RECOVER_URL } from "../../utils/constants";
 import { request, getCookie, setCookie, removeCookie, logError } from "../../utils/functions";
-import { SET_PROFILE_DATA, SET_PROFILE_EDITED } from "./profile";
+import { SET_PROFILE_DATA } from "./profile";
 
 const SET_EXPECTING_RESPONSE = 'SET_EXPECTING_RESPONSE';
 const SET_AUTH_CHECK_STATUS = 'SET_AUTH_CHECK_STATUS';
@@ -117,56 +117,6 @@ const sendLogInRequest = ({ email, password }) => dispatch => {
     .catch(error => logError(error));
 };
 
-const getProfileData = () => dispatch => {
-  dispatch({
-    type: SET_LOGGED_IN_STATUS,
-    status: false
-  });
-  dispatch({
-    type: SET_AUTH_CHECK_STATUS,
-    status: false
-  });
-  dispatch(requestWithToken(PROFILE_DATA_URL, { method: 'GET' }))
-    .then(response => {
-      dispatch({
-        type: SET_PROFILE_DATA,
-        data: response.user
-      });
-      dispatch({
-        type: SET_LOGGED_IN_STATUS,
-        status: true
-      });
-    })
-    .catch(error => logError(error))
-    .finally(() => {
-      dispatch({
-        type: SET_AUTH_CHECK_STATUS,
-        status: true
-      });
-    });
-};
-
-const editProfileData = changes => dispatch => {
-  dispatch(requestWithToken(PROFILE_DATA_URL, {
-    method: 'PATCH',
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify(changes)
-  }))
-    .then(response => {
-      dispatch({
-        type: SET_PROFILE_DATA,
-        data: response.user
-      });
-      dispatch({
-        type: SET_PROFILE_EDITED,
-        status: true
-      })
-    })
-    .catch(error => logError(error));
-};
-
 const sendRecoverRequest = ({ email }) => dispatch => {
   request(RECOVER_URL, {
     method: 'POST',
@@ -231,7 +181,7 @@ const sendLogOurRequest = () => dispatch => {
 };
 
 export {
-  SET_EXPECTING_RESPONSE, SET_AUTH_CHECK_STATUS, SET_LOGGED_IN_STATUS, SET_PROFILE_DATA, SET_FORM_STATUS,
-  sendSignUpRequest, sendLogInRequest, getProfileData, editProfileData, sendRecoverRequest, sendResetRequest,
+  SET_EXPECTING_RESPONSE, SET_AUTH_CHECK_STATUS, SET_LOGGED_IN_STATUS, SET_FORM_STATUS,
+  requestWithToken, sendSignUpRequest, sendLogInRequest, sendRecoverRequest, sendResetRequest,
   sendLogOurRequest
 };
